+++ src/elexmodel/client.py
@@ -28,13 +28,15 @@ class ModelNotEnoughSubunitsException(ModelClientException):
     pass
 
 
-class ModelClient(object):
+class ModelClient:
     """
     Client for generating vote estimates
     """
 
     def __init__(self):
         super().__init__()
+        self.conformalization_data_unit_dict = None
+        self.conformalization_data_agg_dict = None
 
     def _check_input_parameters(
         self,
@@ -48,6 +50,7 @@ class ModelClient(object):
         pi_method,
         beta,
         robust,
+        lambda_,
         handle_unreporting,
     ):
         offices = config_handler.get_offices()
@@ -69,22 +72,54 @@ class ModelClient(object):
         if len(invalid_aggregates) > 0:
             raise ValueError(f"Aggregate(s): {invalid_aggregates} not valid. Please check config")
         model_fixed_effects = config_handler.get_fixed_effects(office)
-        invalid_fixed_effects = [
-            fixed_effect for fixed_effect in fixed_effects if fixed_effect not in model_fixed_effects
-        ]
+        if isinstance(fixed_effects, dict):
+            invalid_fixed_effects = [
+                fixed_effect for fixed_effect in fixed_effects.keys() if fixed_effect not in model_fixed_effects
+            ]
+        else:
+            invalid_fixed_effects = [
+                fixed_effect for fixed_effect in fixed_effects if fixed_effect not in model_fixed_effects
+            ]
         if len(invalid_fixed_effects) > 0:
             raise ValueError(f"Fixed effect(s): {invalid_fixed_effects} not valid. Please check config")
         if pi_method not in {"gaussian", "nonparametric"}:
             raise ValueError(
-                f"Prediction interval method: {pi_method} is not valid. pi_method has to be either `gaussian` or `nonparametric`."
+                f"Prediction interval method: {pi_method} is not valid. \
+                    pi_method has to be either `gaussian` or `nonparametric`."
             )
         if not isinstance(beta, (int, float)):
             raise ValueError("beta is not valid. Has to be either an integer or a float.")
         if not isinstance(robust, bool):
             raise ValueError("robust is not valid. Has to be a boolean.")
+        if not isinstance(lambda_, (float, int)):
+            raise ValueError("lambda is not valid. It has to be numeric.")
+        if lambda_ < 0:
+            raise ValueError("lambda is not valid. It has to be greater than zero.")
         if handle_unreporting not in {"drop", "zero"}:
             raise ValueError("handle_unreporting must be either `drop` or `zero`")
         return True
+
+    def get_all_conformalization_data_unit(self):
+        """
+        This function collects the conformalization data from a model run. It produces a dictionary
+        with two types of data (in the gaussian case): the conformalization points that a
+        distribution is fit to, and the parameters of the resulting guassian distribution.
+        In this unit-level function, the information for one distribution is returned
+        (all units combined). It returns None if get_estimates isn't called, as the
+        values they pull out are generated in that function.
+        """
+        return self.all_conformalization_data_unit_dict
+
+    def get_all_conformalization_data_agg(self):
+        """
+        This function collects the conformalization data from a model run. It produces a dictionary
+        with two types of data (in the gaussian case): the conformalization points that a distribution is
+        fit to, and the parameters of the resulting guassian distribution. In the agg case,
+        distributions for each state (in a multi-state model) are returned. This functions
+        return None if get_estimates isn't called, as the
+        values they pull out are generated in that function.
+        """
+        return self.all_conformalization_data_agg_dict
 
     def get_estimates(
         self,
@@ -111,10 +146,11 @@ class ModelClient(object):
             current_data = pd.DataFrame(current_data[1:], columns=column_values)
         features = kwargs.get("features", [])
         aggregates = kwargs.get("aggregates", ["postal_code", "unit"])
-        fixed_effects = kwargs.get("fixed_effects", [])
+        fixed_effects = kwargs.get("fixed_effects", {})
         pi_method = kwargs.get("pi_method", "nonparametric")
         beta = kwargs.get("beta", 1)
         robust = kwargs.get("robust", False)
+        lambda_ = kwargs.get("lambda_", 0)
         save_output = kwargs.get("save_output", ["results"])
         save_results = "results" in save_output
         save_data = "data" in save_output
@@ -128,7 +164,9 @@ class ModelClient(object):
             "geographic_unit_type": geographic_unit_type,
             "beta": beta,
             "robust": robust,
+            "lambda_": lambda_,
             "features": features,
+            "fixed_effects": fixed_effects,
             "save_conformalization": save_conformalization,
         }
 
@@ -136,7 +174,6 @@ class ModelClient(object):
         config_handler = ConfigHandler(
             election_id, config=raw_config, s3_client=s3.S3JsonUtil(TARGET_BUCKET), save=save_config
         )
-
         self._check_input_parameters(
             config_handler,
             office,
@@ -148,6 +185,7 @@ class ModelClient(object):
             pi_method,
             beta,
             robust,
+            lambda_,
             handle_unreporting,
         )
         states_with_election = config_handler.get_states(office)
@@ -176,7 +214,6 @@ class ModelClient(object):
             current_data,
             estimands,
             geographic_unit_type,
-            fixed_effects=fixed_effects,
             handle_unreporting=handle_unreporting,
         )
 
@@ -189,7 +226,9 @@ class ModelClient(object):
         unexpected_units = data.get_unexpected_units(percent_reporting_threshold, aggregates)
 
         LOG.info(
-            "Model parameters: \n geographic_unit_type: %s, prediction intervals: %s, percent reporting threshold: %s, features: %s, pi_method: %s, aggregates: %s, fixed effects: %s, model settings: %s",
+            "Model parameters: \n geographic_unit_type: %s, prediction intervals: %s, \
+                percent reporting threshold: %s, features: %s, pi_method: %s, aggregates: %s, \
+                fixed effects: %s, model settings: %s",
             geographic_unit_type,
             prediction_intervals,
             percent_reporting_threshold,
@@ -200,7 +239,6 @@ class ModelClient(object):
             model_settings,
         )
 
-        model_settings["expanded_fixed_effects"] = data.expanded_fixed_effects
         if pi_method == "nonparametric":
             model = NonparametricElectionModel(model_settings=model_settings)
         elif pi_method == "gaussian":
@@ -239,16 +277,19 @@ class ModelClient(object):
             aggregates, prediction_intervals, reporting_units, nonreporting_units, unexpected_units
         )
 
+        self.all_conformalization_data_unit_dict = {alpha: {} for alpha in prediction_intervals}
+        self.all_conformalization_data_agg_dict = {alpha: {} for alpha in prediction_intervals}
         for estimand in estimands:
             unit_predictions = model.get_unit_predictions(reporting_units, nonreporting_units, estimand)
             results_handler.add_unit_predictions(estimand, unit_predictions)
             # gets prediciton intervals for each alpha
-            alpha_to_unit_prediction_intervals = {
-                alpha: model.get_unit_prediction_intervals(
+            alpha_to_unit_prediction_intervals = {}
+            for alpha in prediction_intervals:
+                alpha_to_unit_prediction_intervals[alpha] = model.get_unit_prediction_intervals(
                     results_handler.reporting_units, results_handler.nonreporting_units, alpha, estimand
                 )
-                for alpha in prediction_intervals
-            }
+                self.all_conformalization_data_unit_dict[alpha][estimand] = model.get_all_conformalization_data_unit()
+
             results_handler.add_unit_intervals(estimand, alpha_to_unit_prediction_intervals)
 
             for aggregate in results_handler.aggregates:
@@ -260,8 +301,9 @@ class ModelClient(object):
                     aggregate_list,
                     estimand,
                 )
-                alpha_to_agg_prediction_intervals = {
-                    alpha: model.get_aggregate_prediction_intervals(
+                alpha_to_agg_prediction_intervals = {}
+                for alpha in prediction_intervals:
+                    alpha_to_agg_prediction_intervals[alpha] = model.get_aggregate_prediction_intervals(
                         results_handler.reporting_units,
                         results_handler.nonreporting_units,
                         results_handler.unexpected_units,
@@ -269,10 +311,9 @@ class ModelClient(object):
                         alpha,
                         alpha_to_unit_prediction_intervals[alpha].conformalization,
                         estimand,
-                        model_settings,
                     )
-                    for alpha in prediction_intervals
-                }
+                    self.all_conformalization_data_agg_dict[alpha][estimand] = model.get_all_conformalization_data_agg()
+
                 # get all of the prediction intervals here
                 results_handler.add_agg_predictions(
                     estimand, aggregate, estimates_df, alpha_to_agg_prediction_intervals
@@ -365,6 +406,7 @@ class HistoricalModelClient(ModelClient):
             s3_client=s3.S3CsvUtil(TARGET_BUCKET),
             historical=True,
         )
+
         results_to_return = [f"results_{estimand}" for estimand in estimands]
         geo_columns = set(["geographic_unit_fips", "postal_code"] + [a for a in self.aggregates if a != "unit"])
         preprocessed_data = preprocessed_data_handler.data[list(geo_columns) + results_to_return].copy()